Third Amendment of the Constitution of Ireland - 17th June, 1959
The text of the proposed amendment does not appear to be available. The following is the referendum text as presented to the people on 17th June, 1959.

"At present, members of Dáil éireann are elected on a system of proportional representation for constituencies returning at least three members, each voter having a single transferable vote.
It is proposed in the Bill to abolish the system of proportional representation and to adopt, instead, a system of single-member constituencies, each voter having a single non-transferable vote. It is also proposed in the Bill to set up a Commission for the determination and revision of the constituencies, instead of having this done by the Oireachtas, as at present."

By inference, this would have touched on Articles 12 and 16.