Third Amendment of the Consitution of Ireland - 16 October 1968
Again, the text of the proposed amendment does not appear to be available; however, Wikipedia gives it as follows:

The deletion of article 16.2.3, and the insertion in its place of the following:
3. A determination of constituencies shall be so effected that if with respect to each of the constituencies, the number of members to be elected for it is divided into its population (as ascertained at the census immediately[4] preceding the determination) none of the quotients shall be greater, or less, than the average obtained by dividing the total population, as ascertained at the immediately preceding census, by the total number of members of Dáil Éireann[5] by more than one-sixth of that average.
4. A determination of constituencies shall not be effected during a period beginning on the date of a census and ending on the date of the publication of the relevant results (not being provisional results) thereof, and, if the latest time for effecting such a determination falls during such a period and the determination is not effected before the period begins, it shall, notwithstanding anything in this Article, be effected as soon as may be after the period ends.[6]
Subject to the foregoing requirement of this sub-section, regard shall be had at a determination of constituencies to the extent and accessibility of constituencies and the need for securing convenient areas of representation and, subject to those considerations, to the desirability of avoiding the overlapping by constituencies of the boundaries of Administrative Counties (other than boundaries between those Counties and County Boroughs).

and the renumbering of the subsequent articles to "5", "6" and "7".