Third Amendment of the Consitution of Ireland - 10 May 1972
This time for sure. This was the /actual/ Third amendment, after the previous abortive attempts to amend the constitution to remove proprtional representation. This amendment has no relationship to those; its title as Third Amendment is simply numeric.

The legislative act governing this amendment is the Third Amendment of the Constitution Act, 1972 - http://www.irishstatutebook.ie/1972/en/act/cam/0003/print.html